An initial (semi)-working implementation of DES
This commit adds the initial implementation of the block cipher for
DES. Currently, you can only encrypt a single block at a time. The
implementation can encrypt and decrypt messages, but is not in line
with the DES specification, as my BitVector type is currently treating
bit 0 as the least significant bit, while the DES spec (and lookup tables
that I copied from the spec) treat bit 0 as the most significant bit.