Add encryption of ByteStrings
This commit adds functions to the des.hs file that will encrypt an ByteString
of length 8 (i.e. a 64 bit block). This is to make encrypting files easy, as I
can simply read the file into a ByteString and chunk it up.

I also added some QuickCheck properties to ensure that decrypting ciphertext
always returns the plaintext and that my conversion functions from ByteStrings
to BitVectors also have the same property.