Create an executable that encrypts/decrypts files
This commit adds an executable that consumes the DES module to encrypt decrypt files.
It processes the input file, adding padding to ensure that all blocks of the plaintext
are 8 bytes in length. It uses the Crypto.Hash.SHA256 module to generate a key from a
password by hashing the password and taking the first 64 bits to be a key. This likely is
not secure, but it is easy to do for this project. Note that in DES, only 56 of those bits
are considered to be the key, the rest are parity bits.